+++ src/main/java/com/project/taskipro/model/tasks/TaskStatuses.java
@@ -1,0 +1,30 @@
+package com.project.taskipro.model.tasks;
+
+
+import com.project.taskipro.model.tasks.enums.StatusType;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name="task_statuses")
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class TaskStatuses {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column(name = "status_type")
+    @Enumerated(EnumType.STRING)
+    private StatusType statusType;
+
+    @ManyToOne
+    @JoinColumn(name = "task_id")
+    private Tasks task;
+}
