+++ src/main/java/com/project/taskipro/model/tasks/enums/StatusType.java
@@ -1,4 +1,4 @@
-package com.project.taskipro.entity.tasks;
+package com.project.taskipro.model.tasks.enums;
 
 public enum StatusType {
     BACKLOG,
