+++ src/main/java/com/project/taskipro/model/user/UserSubscription.java
@@ -1,4 +1,4 @@
-package com.project.taskipro.entity.user;
+package com.project.taskipro.model.user;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.util.Date;
-import java.util.List;
 
 @Data
 @AllArgsConstructor
