+++ src/main/java/com/project/taskipro/repository/TokenRepository.java
@@ -1,6 +1,6 @@
 package com.project.taskipro.repository;
 
-import com.project.taskipro.entity.auth.Token;
+import com.project.taskipro.model.auth.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
