+++ src/main/java/com/project/taskipro/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.project.taskipro.repository;
 
-import com.project.taskipro.entity.user.User;
+import com.project.taskipro.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
