+++ src/main/java/com/project/taskipro/service/JwtService.java
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
-import com.project.taskipro.entity.user.User;
+import com.project.taskipro.model.user.User;
 
 import javax.crypto.SecretKey;
 import java.util.Date;
