+++ src/main/java/com/project/taskipro/service/UserServiceImpl.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import com.project.taskipro.entity.user.User;
+import com.project.taskipro.model.user.User;
 
 @RequiredArgsConstructor
 @Service
