+++ src/main/resources/application.yml
@@ -9,8 +9,11 @@ spring:
     driverClassName: org.postgresql.Driver
   jpa:
     hibernate:
-      ddl-auto: create-drop
+      ddl-auto: none
     open-in-view: true
+  liquibase:
+    enabled: true
+
 security:
     jwt:
       secret_key: 5355925dd8425f96b24d4ef73f5fb56916ab8cd03e9bb07d333fcb30b304f13e90dd4d7e6447ade7b11dcd7bd7eae4042537c4e73e7ef96dc5340669ced406b3feb79579bf15718167fc27752acf4baf8b0bbad3a6b7ae4d28b9eba9d1326b897605714c9beddf6b521f7645ffdf255278995a2f496474cb7fff0bdcfdd85ea1
