+++ src/main/resources/db/changelog/changeset/create-subscription-table.yaml
@@ -1,0 +1,64 @@
+databaseChangeLog:
+  - changeSet:
+      id: 2
+      author: Dima
+      changes:
+        - createTable:
+            tableName: subscription_type
+            columns:
+              - column:
+                  name: id
+                  type: BIGINT
+                  autoIncrement: true
+                  constraints:
+                    primaryKey: true
+                    nullable: false
+              - column:
+                  name: subscription_type
+                  type: VARCHAR(55)
+                  constraints:
+                    nullable: false
+              - column:
+                  name: desk_limit
+                  type: INT
+                  constraints:
+                    nullable: false
+              - column:
+                  name: days_limit
+                  type: INT
+                  constraints:
+                    nullable: false
+              - column:
+                  name: price
+                  type: FLOAT
+                  constraints:
+                    nullable: false
+        - createTable:
+            tableName: user_subscription
+            columns:
+              - column:
+                  name: id_user_subscription
+                  type: BIGINT
+                  autoIncrement: true
+                  constraints:
+                    primaryKey: true
+                    nullable: false
+              - column:
+                  name: subscription_start_date
+                  type: DATE
+                  constraints:
+                    nullable: false
+              - column:
+                  name: user_id
+                  type: BIGINT
+                  constraints:
+                    nullable: false
+                    foreignKeyName: fk_subscription_user_id
+                    references: users(id)
+              - column:
+                  name: subscription_type_id
+                  type: BIGINT
+                  constraints:
+                    nullable: false
+                    foreignKeyName: fk_subscription_type_id
+                    references: subscription_type(id)
