+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,0 +1,9 @@
+databaseChangeLog:
+  - include:
+      file: db/changelog/changeset/create-user-tables.yaml
+  - include:
+      file: db/changelog/changeset/create-subscription-table.yaml
+  - include:
+      file: db/changelog/changeset/create-desks-tables.yaml
+  - include:
+      file: db/changelog/changeset/create-tasks-tables.yaml
